microservices/user_management/src/requirements.txt
sqlalchemy
python-dotenv
PyYAML
pandas
pydantic-settings 
httpx
Tenacity
bcrypt==4.0.1
passlib[bcrypt]
python-jose[cryptography]
python-multipart